Clean-up multiselect code
Allocate memory for the multiselect buffer in chunks, similar to how the
`choices` buffer is resized. This results in potentially more unused
(i.e. wasted) memory, but in fewer invocations of `malloc`/`realloc`,
which is probably more important, considering the size of the
multiselect array will never be that large.